Added fix PHPStorm folder permissions task
When phpstorm needs to update patches it needs that the user running phpstorm
have permissions on the installation folder, so I added this task to change the
owner of these files.
Also a register was added to the unpack task so it will only run if is phpstorm
is not installed.